fixed overwrite bug in pn_quote_data
git-svn-id: https://svn.apache.org/repos/asf/qpid/proton/trunk@1356738 13f79535-47bb-0310-9956-ffa450edef68

+++ proton-c/src/util.c
@@ -37,13 +37,13 @@ ssize_t pn_quote_data(char *dst, size_t capacity, const char *src, size_t size)
   {
     uint8_t c = src[i];
     if (isprint(c)) {
-      if (idx < capacity - 1) {
+      if (idx < (int) (capacity - 1)) {
         dst[idx++] = c;
       } else {
         return PN_OVERFLOW;
       }
     } else {
-      if (idx < capacity - 4) {
+      if (idx < (int) (capacity - 4)) {
         idx += sprintf(dst + idx, "\\x%.2x", c);
       } else {
         return PN_OVERFLOW;
